refactor(cli): using entry path to prod entry path
