--- packages/cli/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Document</title>
-    </head>
-    <body>
-        <div id="root">
-            <!--ssr-outlet-->
-        </div>
-        <script type="module" src="./src/ssr/entry-client.tsx"></script>
-    </body>
-</html>
