+++ packages/cli/src/env.d.ts
@@ -1,0 +1,7 @@
+interface ImportMetaEnv extends Readonly<Record<string, string>> {
+  readonly ESBUILD_PROD: boolean;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}
