+++ packages/cli/src/ssr/createServer.ts
@@ -8,29 +8,31 @@ export const createServer = async () => {
   const vite = await createViteServer({
     server: { middlewareMode: true },
     appType: "custom",
-    resolve: {
-      alias: {
-        "@assetbox/manager": path.resolve(
-          __dirname,
-          "../../manager/dist/entry.mjs"
-        ),
-      },
-    },
+    // resolve: {
+    //   alias: {
+    //     "@assetbox/manager": path.resolve(
+    //       process.cwd(),
+    //       "../../manager/dist/entry.mjs"
+    //     ),
+    //   },
+    // },
   });
 
   app.use(vite.middlewares);
   app.use("*", async (req, res, next) => {
     const url = req.originalUrl;
-
     try {
       let template = fs.readFileSync(
-        path.resolve(__dirname, "../index.html"),
+        path.resolve(process.cwd(), "templates", "index.html"),
         "utf-8"
       );
       template = await vite.transformIndexHtml(url, template);
-      const { render } = await vite.ssrLoadModule(
-        path.resolve(__dirname, "../src/ssr/entry-server.tsx")
-      );
+
+      const entryServerModulePath = import.meta.env.ESBUILD_PROD
+        ? path.resolve(process.cwd(), "ssr", "entryServer.cjs")
+        : path.resolve(process.cwd(), "entryServer.tsx");
+
+      const { render } = await vite.ssrLoadModule(entryServerModulePath);
       const appHtml = await render(url);
       const html = template.replace(`<!--ssr-outlet-->`, appHtml);
