+++ packages/cli/src/ssr/templates/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+  </head>
+  <body>
+    <div id="root">
+      <!--ssr-outlet-->
+    </div>
+    <script type="module" src="./entry-client.tsx"></script>
+  </body>
+</html>
